Implement missing importers
This commit implements the missing CSV importers, which were left out as
they would require PATCH requests. To support this new kind of request,
"private" keys are added to the request body object, to be filtered out
and used to change the request.

It also corrects errors that were preventing the script from running,
and institutions from being imported.